+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 - Fix warning when compression options isn't standard size
 - In `from_reader(..)`, show info about flags used
+- Add `Filesystem::from_reader(..)` and `Filesystem::from_reader_with_offset(..)`
+  which calls `Squashfs::from_reader(..)` and `Squashfs::from_reader_with_offset(..)` and `Squashfs::into_filesystem(..)`.
 
 ## [v0.4.0] - 2021-01-04
 - Add `mod_time` from `Squashfs` to `Filesystem` used in creation of new image with `to_bytes(..)`